+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.21.2...HEAD) - TBD
 
+### :bug: Fixed
+
+- False positive `negative_data_rejection` when using `@schemathesis.auth` with a cookie security scheme. [#4236](https://github.com/schemathesis/schemathesis/issues/4236)
+
 ## [4.21.2](https://github.com/schemathesis/schemathesis/compare/v4.21.1...v4.21.2) - 2026-06-09
 
 ### :bug: Fixed